Move virtual path crawler tool method to ServiceTools.
